make monarq.sim accessible + refactoring
