+++ pennylane_calculquebec/device_exception.py
@@ -1,0 +1,6 @@
+
+
+class DeviceException(Exception):
+    def __init__(self, message):
+        super().__init__(message)
+
